Modularização do código.
